# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: سند – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Update user status (ban/activate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /admin/providers/{id}/verify:
    patch:
      operationId: verifyProvider
      tags: [admin]
      summary: Verify/reject provider account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderVerifyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── ADMIN ───────────────────────────────────────────────────────────────────
...
    AdminUsersPage:
      type: object
      required: [users, total, page]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
    AdminUser:
      type: object
      required: [id, name, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        completedJobs:
          type: ["integer", "null"]
        isVerified:
          type: ["boolean", "null"]
        createdAt:
          type: string
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, banned, pending]
    ProviderVerifyInput:
      type: object
      required: [isVerified]
      properties:
        isVerified:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
...